line chart: refactor to move interface for Scale to types file (#4289)
Moved the Scale interface definition to the type so dependents do
not have to load the implementation and keep it private to the lib.
