+++ tensorboard/webapp/widgets/line_chart_v2/lib/scale_test.ts
@@ -13,8 +13,8 @@ See the License for the specific language governing permissions and
 limitations under the License.
 ==============================================================================*/
 
-import {Scale, createScale} from './scale';
-import {ScaleType} from './scale_types';
+import {createScale} from './scale';
+import {Scale, ScaleType} from './scale_types';
 
 describe('line_chart_v2/lib/scale test', () => {
   describe('linear', () => {
